Added Json transformer utility
